// src/api/dataset/models/dataset-chunk-process.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 决定proto引用路径和rpc路由
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: version not set
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { DatasetCompressType } from './dataset-compress-type';
 /**
 * 
 *
 * @export
 * @interface DatasetChunkProcess
 */
export interface DatasetChunkProcess {

    /**
     * @type {number}
     * @memberof DatasetChunkProcess
     */
    storeType?: number;

    /**
     * @type {DatasetCompressType}
     * @memberof DatasetChunkProcess
     */
    compressType?: DatasetCompressType;
}