// src/api/dataset/models/dataset-dataset-extend.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 决定proto引用路径和rpc路由
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: version not set
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { DatasetChunkProcess } from "./dataset-chunk-process";
import { DatasetDataProcess } from "./dataset-data-process";
import { DatasetValueProcess } from "./dataset-value-process";
/**
 *
 *
 * @export
 * @interface DatasetDatasetExtend
 */
export interface DatasetDatasetExtend {
  /**
   * @type {DatasetDataProcess}
   * @memberof DatasetDatasetExtend
   */
  dataProcess?: DatasetDataProcess;

  /**
   * @type {DatasetChunkProcess}
   * @memberof DatasetDatasetExtend
   */
  chunkProcess?: DatasetChunkProcess;

  /**
   * @type {DatasetValueProcess}
   * @memberof DatasetDatasetExtend
   */
  valueProcess?: DatasetValueProcess;
}