// src/api/dataset/models/dataset-dataset-info-a.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 决定proto引用路径和rpc路由
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: version not set
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { DatasetDatasetExtend } from './dataset-dataset-extend';
import { DatasetOpInfoA } from './dataset-op-info-a';
import { DatasetStatus } from './dataset-status';
 /**
 * 
 *
 * @export
 * @interface DatasetDatasetInfoA
 */
export interface DatasetDatasetInfoA {

    /**
     * @type {string}
     * @memberof DatasetDatasetInfoA
     */
    datasetId?: string;

    /**
     * @type {string}
     * @memberof DatasetDatasetInfoA
     */
    datasetName?: string;

    /**
     * @type {string}
     * @memberof DatasetDatasetInfoA
     */
    remark?: string;

    /**
     * @type {DatasetDatasetExtend}
     * @memberof DatasetDatasetInfoA
     */
    datasetExtend?: DatasetDatasetExtend;

    /**
     * @type {string}
     * @memberof DatasetDatasetInfoA
     */
    valueTotal?: string;

    /**
     * @type {string}
     * @memberof DatasetDatasetInfoA
     */
    createTime?: string;

    /**
     * @type {string}
     * @memberof DatasetDatasetInfoA
     */
    processedTime?: string;

    /**
     * @type {DatasetOpInfoA}
     * @memberof DatasetDatasetInfoA
     */
    op?: DatasetOpInfoA;

    /**
     * @type {DatasetStatus}
     * @memberof DatasetDatasetInfoA
     */
    status?: DatasetStatus;

    /**
     * @type {string}
     * @memberof DatasetDatasetInfoA
     */
    statusInfo?: string;

    /**
     * @type {string}
     * @memberof DatasetDatasetInfoA
     */
    activateTime?: string;
}