// src/api/dataset/models/dataset-data-source-uri-file.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 决定proto引用路径和rpc路由
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: version not set
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { DatasetKV } from "./dataset-kv";
/**
 *
 *
 * @export
 * @interface DatasetDataSourceUriFile
 */
export interface DatasetDataSourceUriFile {
  /**
   * @type {string}
   * @memberof DatasetDataSourceUriFile
   */
  uri?: string;

  /**
   * @type {Array<DatasetKV>}
   * @memberof DatasetDataSourceUriFile
   */
  headers?: Array<DatasetKV>;

  /**
   * @type {boolean}
   * @memberof DatasetDataSourceUriFile
   */
  insecureSkipVerify?: boolean;

  /**
   * @type {string}
   * @memberof DatasetDataSourceUriFile
   */
  proxy?: string;

  /**
   * @type {string}
   * @memberof DatasetDataSourceUriFile
   */
  method?: string;
}